requirements.txt
astroid==1.6.5
boto3==1.7.52
botocore==1.10.52
cached-property==1.4.3
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
django-cors-headers==2.3.0
django-filter==1.1.0
django-quiet-runserver==0.2.0
django-storages==1.6.6
django-url-filter==0.3.5
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
djangorestframework-gis==0.13
djangorestframework-queryfields==1.0.0
djangorestframework-recursive==0.1.2
docutils==0.14
drf-nested-routers==0.90.2
enum-compat==0.0.2
enum34==1.1.6
idna==2.7
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
pdfkit==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pylint==1.9.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
wrapt==1.10.11
